+++ packages/page-docs/src/scripts/layout/Landing.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { Footer } from './footer';
-import { Header } from './header';
+import { Footer } from './Footer';
+import { Header } from './Header';
 import { TagLine } from '../components';
 
 export const Landing: React.SFC = ({ children }) => (