+++ packages/piral-cli/src/apps/build-piral.ts
@@ -1,6 +1,6 @@
 import * as Bundler from 'parcel-bundler';
 import { join, dirname, basename } from 'path';
-import { extendConfig } from './common';
+import { extendConfig, setStandardEnvs } from './common';
 
 export interface BuildPiralOptions {
   entry?: string;
@@ -12,25 +12,25 @@ export const buildPiralDefaults = {
   target: './dist/index.html',
 };
 
-export function buildPiral(baseDir = process.cwd(), options: BuildPiralOptions = {}) {
+export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOptions = {}) {
   const { entry = buildPiralDefaults.entry, target = buildPiralDefaults.target } = options;
   const entryFiles = join(baseDir, entry);
 
-  process.env.NODE_ENV = 'production';
+  setStandardEnvs({
+    production: true,
+    target: dirname(entry),
+  });
+  const bundler = new Bundler(
+    entryFiles,
+    extendConfig({
+      outDir: dirname(target),
+      outFile: basename(target),
+      watch: false,
+      minify: true,
+      scopeHoist: false,
+      contentHash: true,
+    }),
+  );
 
-  (async function() {
-    const bundler = new Bundler(
-      entryFiles,
-      extendConfig({
-        outDir: dirname(target),
-        outFile: basename(target),
-        watch: false,
-        minify: true,
-        scopeHoist: false,
-        contentHash: true,
-      }),
-    );
-
-    await bundler.bundle();
-  })();
+  await bundler.bundle();
 }